new file:   Apps/JSON/JSON.zip 	new file:   Apps/LearningWeb/Learning-Web.zip 	new file:   Apps/PyKorean/PyKorean.zip 	new file:   Data/InstallLog.txt 	modified:   Program.json 	modified:   Scripts/Install.py 	modified:   Scripts/JSONMaker.py 	renamed:    Scripts/HTMLQuiz.py -> Scripts/LearningWeb.py 	modified:   Scripts/PyKorean.py 	modified:   Scripts/Short.py 	modified:   Scripts/__init__.py 	modified:   Unzip.sh

+++ Unzip.sh
@@ -1,6 +1,23 @@
-# Going to the folder that has the JSON.zip file, and unzipping the zip file.
-cd $1 && unzip -q JSON.zip && rm JSON.zip
-# Going to the folder that has the PyKorean.zip file, and unzipping the zip file.
-cd $2 && unzip -q PyKorean.zip && rm PyKorean.zip
-# Going to the folder that has the HTMLQuiz.zip file, and unzipping the zip file.
-cd $3 && unzip -q HTMLQuiz.zip && rm HTMLQuiz.zip
+# Checking the option for installing All
+if [ $4 == "y" ]; then
+    # Going to the folder that has the JSON.zip file, and unzipping the zip file.
+    cd $1 && unzip -q JSON.zip && rm JSON.zip
+    # Going to the folder that has the PyKorean.zip file, and unzipping the zip file.
+    cd $2 && unzip -q PyKorean.zip && rm PyKorean.zip
+    # Going to the folder that has the HTMLQuiz.zip file, and unzipping the zip file.
+    cd $3 && unzip -q Learning-Web.zip && rm Learning-Web.zip
+else
+    # Checking JSON.
+    if [ $5 -eq "y"]; then
+        # Going to the folder that has the JSON.zip file, and unzipping the zip file.
+        cd $1 && unzip -q JSON.zip && rm JSON.zip
+    fi
+    if [ $6 -eq "y" ]; then
+        # Going to the folder that has the PyKorean.zip file, and unzipping the zip file.
+        cd $2 && unzip -q PyKorean.zip && rm PyKorean.zip
+    fi
+    if [ $7 -eq "y" ]; then
+        # Going to the folder that has the HTMLQuiz.zip file, and unzipping the zip file.
+        cd $3 && unzip -q Learning-Web.zip && rm Learning-Web.zip
+    fi
+fi
